backend/requirements.txt
# Phase II requirements
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.8
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.3
python-multipart==0.0.6
alembic==1.12.1
asyncpg==0.29.0
python-dotenv==1.0.0
emails==0.6
itsdangerous==2.1.2
jinja2==3.1.2
cryptography==41.0.7
python-dateutil==2.8.2
python-slugify==8.0.1
Pillow==10.1.0
phonenumbers==8.13.25
email-validator==2.1.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
flower==2.0.1
gunicorn==21.2.0
whitenoise==6.6.0
uvloop==0.19.0
httptools==0.6.1
watchfiles==0.21.0
anyio==4.0.0
typing-extensions==4.8.0
idna==3.4
certifi==2023.7.22
charset-normalizer==3.3.2
urllib3==2.0.7
requests==2.31.0
# Phase III requirements
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.5
langchain-community>=0.0.19